Move management of user formats into DocIndexerFactoryUserFormats
It now is in control of creating and destroying user formats,
including deleting the disk files.

- Add lazy-loading of other user's formats. When a new corpus is created
using an unknown format, or when a format refers to it using
"baseFormat" or "inputFormat" in the linkedDocuments section, the
userformats factory will attempt to locate and load it.
- Disabled threaded indexing for class based indexers until some
concurrency issues are solved with them (specifically
outputPrefixMapping in DocIndexerXmlHandlers)
- No longer indicate simple parsing exceptions as IO errors in
InputFormatReader.
- Sanitize usernames when attempting to use another user's formats.